# gui/app/openapi/api/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "FastAPI", "version": "0.1.0" },
	"servers": [{ "url": "/api" }],
	"paths": {
...
		"/v1/items/{item_id}": {
			"get": {
				"summary": "Get Item",
				"description": "Get item.",
				"operationId": "get_item_v1_items__item_id__get",
				"parameters": [
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the item"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ItemRead" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
		"/v1/me/items/{item_id}": {
			"delete": {
				"summary": "Delete Client Item",
				"description": "Delete the specified item owned by the client.",
				"operationId": "delete_client_item_v1_me_items__item_id__delete",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the item."
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"get": {
				"summary": "Get Client Item By Id",
				"description": "Get client's item by id.",
				"operationId": "get_client_item_by_id_v1_me_items__item_id__get",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the item."
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ItemRead" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"post": {
				"summary": "Update Client Item",
				"description": "Update client's item.",
				"operationId": "update_client_item_v1_me_items__item_id__post",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the item."
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ItemUpdate",
								"title": "Item fields to update."
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ItemPrivateRead" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
		"/v1/me/chats/{chat_id}/messages": {
			"post": {
				"summary": "Send Message To Chat",
				"description": "Send message to chat.",
				"operationId": "send_message_to_chat_v1_me_chats__chat_id__messages_post",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "chat_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "\\d+-\\d+",
							"title": "The ID of the chat."
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/ChatMessageCreate" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ChatMessageRead" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"get": {
				"summary": "List Client Chat Messages",
				"description": "List messages in the chat.",
				"operationId": "list_client_chat_messages_v1_me_chats__chat_id__messages_get",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "chat_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"pattern": "\\d+-\\d+",
							"title": "The ID of the chat."
						}
					},
					{
						"name": "seen",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "boolean" }, { "type": "null" }],
							"title": "Seen",
							"description": "Only select messages that have been seen.",
							"examples": [true, false]
						},
						"description": "Only select messages that have been seen."
					},
					{
						"name": "n",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{ "type": "integer", "maximum": 128, "exclusiveMinimum": 0 },
								{ "type": "null" }
							],
							"title": "Limit returned messages count",
							"description": "Limit the number of messages returned.",
							"examples": [[42]],
							"default": 64
						},
						"description": "Limit the number of messages returned."
					},
					{
						"name": "cid",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{ "type": "integer", "minimum": 0 },
								{ "type": "null" }
							],
							"title": "Page cursor for chat message ID"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/ChatMessageRead" },
									"title": "Response List Client Chat Messages V1 Me Chats  Chat Id  Messages Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
		"/v1/users/{user_id}/items/{item_id}": {
			"get": {
				"summary": "Get Client Item By Id",
				"description": "Get user's item by id.",
				"operationId": "get_client_item_by_id_v1_users__user_id__items__item_id__get",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the user."
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 0,
							"title": "The ID of the item"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ItemRead" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ChatId": { "type": "string" },
			"ChatMessageCreate": {
				"properties": { "text": { "type": "string", "title": "Text" } },
				"additionalProperties": false,
				"type": "object",
				"required": ["text"],
				"title": "ChatMessageCreate"
			},
			"ChatMessageRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"chat_id": { "$ref": "#/components/schemas/ChatId" },
					"message_type": { "$ref": "#/components/schemas/ChatMessageType" },
					"sender_id": {
						"anyOf": [{ "type": "integer" }, { "type": "null" }],
						"title": "Sender Id"
					},
					"creation_date": {
						"type": "string",
						"format": "date-time",
						"title": "Creation Date"
					},
					"seen": { "type": "boolean", "title": "Seen" },
					"text": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Text"
					},
					"loan_request_id": {
						"anyOf": [{ "type": "integer" }, { "type": "null" }],
						"title": "Loan Request Id"
					},
					"loan_id": {
						"anyOf": [{ "type": "integer" }, { "type": "null" }],
						"title": "Loan Id"
					},
					"item_id": { "type": "integer", "title": "Item Id" },
					"borrower_id": { "type": "integer", "title": "Borrower Id" }
				},
				"type": "object",
				"required": [
					"id",
					"chat_id",
					"message_type",
					"sender_id",
					"creation_date",
					"seen",
					"text",
					"loan_request_id",
					"loan_id",
					"item_id",
					"borrower_id"
				],
				"title": "ChatMessageRead"
			},
			"ChatMessageType": {
				"type": "integer",
				"enum": [1, 2, 3, 4, 5, 6, 7, 8, 9],
				"title": "ChatMessageType",
				"x-enum-varnames": [
					"text",
					"loan_request_created",
					"loan_request_cancelled",
					"loan_request_accepted",
					"loan_request_rejected",
					"loan_started",
					"loan_ended",
					"item_not_available",
					"item_available"
				]
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ItemPreviewRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"name": { "type": "string", "title": "Name" },
					"targeted_age_months": {
						"prefixItems": [
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] },
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] }
						],
						"type": "array",
						"maxItems": 2,
						"minItems": 2,
						"title": "Targeted Age Months"
					},
					"first_image_name": { "type": "string", "title": "First Image Name" },
					"available": { "type": "boolean", "title": "Available" },
					"owner_id": { "type": "integer", "title": "Owner Id" }
				},
				"type": "object",
				"required": [
					"id",
					"name",
					"targeted_age_months",
					"first_image_name",
					"available",
					"owner_id"
				],
				"title": "ItemPreviewRead"
			},
			"ItemPrivateRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"name": { "type": "string", "title": "Name" },
					"description": { "type": "string", "title": "Description" },
					"targeted_age_months": {
						"prefixItems": [
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] },
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] }
						],
						"type": "array",
						"maxItems": 2,
						"minItems": 2,
						"title": "Targeted Age Months"
					},
					"images_names": {
						"items": { "type": "string" },
						"type": "array",
						"title": "Images Names"
					},
					"available": { "type": "boolean", "title": "Available" },
					"owner_id": { "type": "integer", "title": "Owner Id" },
					"owner": { "$ref": "#/components/schemas/UserPreviewRead" },
					"regions": {
						"items": { "$ref": "#/components/schemas/RegionRead" },
						"type": "array",
						"title": "Regions"
					},
					"likes_count": { "type": "integer", "title": "Likes Count" },
					"blocked": { "type": "boolean", "title": "Blocked" },
					"loans": {
						"items": { "$ref": "#/components/schemas/LoanRead" },
						"type": "array",
						"title": "Loans"
					}
				},
				"type": "object",
				"required": [
					"id",
					"name",
					"description",
					"targeted_age_months",
					"images_names",
					"available",
					"owner_id",
					"owner",
					"regions",
					"likes_count",
					"blocked",
					"loans"
				],
				"title": "ItemPrivateRead"
			},
...
			"ItemRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"name": { "type": "string", "title": "Name" },
					"description": { "type": "string", "title": "Description" },
					"targeted_age_months": {
						"prefixItems": [
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] },
							{ "anyOf": [{ "type": "integer" }, { "type": "null" }] }
						],
						"type": "array",
						"maxItems": 2,
						"minItems": 2,
						"title": "Targeted Age Months"
					},
					"images_names": {
						"items": { "type": "string" },
						"type": "array",
						"title": "Images Names"
					},
					"available": { "type": "boolean", "title": "Available" },
					"owner_id": { "type": "integer", "title": "Owner Id" },
					"owner": { "$ref": "#/components/schemas/UserPreviewRead" },
					"regions": {
						"items": { "$ref": "#/components/schemas/RegionRead" },
						"type": "array",
						"title": "Regions"
					},
					"likes_count": { "type": "integer", "title": "Likes Count" }
				},
				"type": "object",
				"required": [
					"id",
					"name",
					"description",
					"targeted_age_months",
					"images_names",
					"available",
					"owner_id",
					"owner",
					"regions",
					"likes_count"
				],
				"title": "ItemRead"
			},
			"ItemUpdate": {
				"properties": {
					"name": {
						"anyOf": [
							{ "type": "string", "maxLength": 30, "minLength": 5 },
							{ "type": "null" }
						],
						"title": "Name"
					},
					"description": {
						"anyOf": [
							{ "type": "string", "maxLength": 600, "minLength": 20 },
							{ "type": "null" }
						],
						"title": "Description"
					},
					"images": {
						"anyOf": [
							{ "items": { "type": "string" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Images"
					},
					"targeted_age_months": {
						"anyOf": [
							{
								"prefixItems": [
									{ "anyOf": [{ "type": "integer" }, { "type": "null" }] },
									{ "anyOf": [{ "type": "integer" }, { "type": "null" }] }
								],
								"type": "array",
								"maxItems": 2,
								"minItems": 2
							},
							{ "type": "null" }
						],
						"title": "Targeted Age Months"
					},
					"regions": {
						"anyOf": [
							{ "items": { "type": "integer" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Regions"
					},
					"blocked": {
						"anyOf": [{ "type": "boolean" }, { "type": "null" }],
						"title": "Blocked"
					}
				},
				"additionalProperties": false,
				"type": "object",
				"title": "ItemUpdate"
			},
			"LoanRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"item": { "$ref": "#/components/schemas/ItemPreviewRead" },
					"owner": { "$ref": "#/components/schemas/UserPreviewRead" },
					"borrower": { "$ref": "#/components/schemas/UserPreviewRead" },
					"chat_id": { "$ref": "#/components/schemas/ChatId" },
					"during": {
						"prefixItems": [
							{
								"anyOf": [
									{ "type": "string", "format": "date-time" },
									{ "type": "null" }
								]
							},
							{
								"anyOf": [
									{ "type": "string", "format": "date-time" },
									{ "type": "null" }
								]
							}
						],
						"type": "array",
						"maxItems": 2,
						"minItems": 2,
						"title": "During"
					},
					"active": { "type": "boolean", "title": "Active" }
				},
				"type": "object",
				"required": [
					"id",
					"item",
					"owner",
					"borrower",
					"chat_id",
					"during",
					"active"
				],
				"title": "LoanRead"
			},
...
			"RegionRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"name": { "type": "string", "title": "Name" }
				},
				"type": "object",
				"required": ["id", "name"],
				"title": "RegionRead"
			},
...
			"UserPreviewRead": {
				"properties": {
					"id": { "type": "integer", "title": "Id" },
					"name": { "type": "string", "title": "Name" },
					"avatar_seed": { "type": "string", "title": "Avatar Seed" }
				},
				"type": "object",
				"required": ["id", "name", "avatar_seed"],
				"title": "UserPreviewRead"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location"
					},
					"msg": { "type": "string", "title": "Message" },
					"type": { "type": "string", "title": "Error Type" }
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": { "scopes": {}, "tokenUrl": "/api/v1/auth/login" }
				}
			}
		}
	}
}